Refactor API calls: Replace axios with fetch for torrent and state retrieval with token-based authentication
