+++ package.json
@@ -29,7 +29,6 @@
     "@tanstack/react-query": "^5.74.4",
     "@tanstack/react-table": "^8.21.3",
     "@types/byte-size": "^8.1.2",
-    "axios": "^1.9.0",
     "byte-size": "^9.0.1",
     "change-case": "^5.4.4",
     "class-variance-authority": "^0.7.1",
