+++ src/lib/api.ts
@@ -1,23 +1,77 @@
 'use client';
 
+import {
+  type ApiResponse,
+  NormalizedTorrentForTable,
+  TorrentSpeedForTable,
+} from '@/types';
 import 'client-only';
 
-import axios from 'axios';
+export async function fetchAllTorrents(
+  token?: string | null,
+): Promise<Record<string, NormalizedTorrentForTable> | undefined> {
+  const response = await fetch('/api/deluge/torrents', {
+    method: 'GET',
+    headers: {
+      ...(token && { Authorization: `Bearer ${token}` }),
+    },
+  });
 
-import { useDelugeListStore } from '@/lib/store';
+  const result = (await response.json()) as ApiResponse<
+    Record<string, NormalizedTorrentForTable>
+  >;
 
-const delugeNextBaseUrl = useDelugeListStore.getState().delugeNextBaseUrl;
+  if (!result.success) throw new Error(result.error);
+  return result.data;
+}
 
-export const api = axios.create({
-  baseURL: `${delugeNextBaseUrl}/api/deluge`,
-  timeout: 10_000,
-});
+export async function fetchAllStates(
+  token?: string | null,
+): Promise<string[] | undefined> {
+  const response = await fetch('/api/deluge/states', {
+    method: 'GET',
+    headers: {
+      ...(token && { Authorization: `Bearer ${token}` }),
+    },
+  });
 
-// Optional: request/response interceptors for error handling, login, logging, etc.
-api.interceptors.response.use(
-  (response) => response,
-  (error) => {
-    // normalize error
-    return Promise.reject(error.response?.data?.error || error.message);
-  },
-);
+  const result = (await response.json()) as ApiResponse<string[]>;
+
+  if (!result.success) throw new Error(result.error);
+  return result.data;
+}
+
+export async function fetchActiveTorrentsSpeed(
+  token?: string | null,
+): Promise<Record<string, TorrentSpeedForTable> | undefined> {
+  const response = await fetch('/api/deluge/torrents/active/speed', {
+    method: 'GET',
+    headers: {
+      ...(token && { Authorization: `Bearer ${token}` }),
+    },
+  });
+
+  const result = (await response.json()) as ApiResponse<
+    Record<string, TorrentSpeedForTable>
+  >;
+
+  if (!result.success) throw new Error(result.error);
+  return result.data;
+}
+
+export async function fetchAllLabels(
+  token?: string | null,
+): Promise<string[] | undefined> {
+  const response = await fetch('/api/deluge/labels', {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error('Failed to fetch labels');
+  }
+  const data = (await response.json()) as ApiResponse<string[]>;
+  return data.data;
+}
